+++ docs/source/en/index.mdx
@@ -288,6 +288,7 @@ Flax), PyTorch, and/or TensorFlow.
 |            Marian             |       ✅       |       ❌       |       ✅        |         ✅         |      ✅      |
 |           MarkupLM            |       ✅       |       ✅       |       ✅        |         ❌         |      ❌      |
 |          MaskFormer           |       ❌       |       ❌       |       ✅        |         ❌         |      ❌      |
+|        MaskFormerSwin         |       ❌       |       ❌       |       ❌        |         ❌         |      ❌      |
 |             mBART             |       ✅       |       ✅       |       ✅        |         ✅         |      ✅      |
 |         Megatron-BERT         |       ❌       |       ❌       |       ✅        |         ❌         |      ❌      |
 |          MobileBERT           |       ✅       |       ✅       |       ✅        |         ✅         |      ❌      |
